refactor: 스토리북 폴더 구조 변경 - 페이지 폴더 내에 stories 폴더를 만들어서 스토리북 파일을 모았습니다.
